pyproject.toml
[project]
name = "monopoly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "bs4>=0.0.2",
    "fastapi>=0.118.0",
    "pillow>=12.0.0",
    "pydantic[email]>=2.11.10",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
